+++ poketechnika/Bar.h
@@ -4,13 +4,13 @@
 class Bar : public Drawable {
 private:
 	int max_value;
-	int value;
+	float value;
 public:
 	static sf::Font _font;
 	sf::Text text;
 	Drawable progress;
 
-	Bar(int max_val, int val, const sf::Vector2f pos, 
+	Bar(int max_val, float val, const sf::Vector2f pos, 
 		const std::string& texture_file = "assets/textures/bar_background.png"
 	);
 	static void load_font(const std::string& font_file = "assets/Roboto-Bold.ttf");
